# public/openapi.yaml
openapi: 3.0.3
info:
  title: Aave Incentives API
  description: |
    API to fetch all incentives available on the Aave V3 protocol across multiple chains.

    This API aggregates incentives from multiple sources:
    - ACI MASIv Infra incentives
    - Merkl infra incentives
    - Onchain incentives
    - Local config incentives

  version: 1.0.0
  contact:
    name: ACI
    url: https://www.aavechan.com/
  license:
    name: MIT
servers:
  - url: https://aave-incentives-api.vercel.app/
    description: Production server
  - url: http://localhost:5050
    description: Local development
...
paths:
  /incentives:
    get:
      tags:
        - Incentives
      summary: Get all incentives
      description: |
        Fetches all available incentives with optional filtering, sorting, and pagination.

        Results are cached for 5 minutes for optimal performance.
      operationId: getIncentives
      parameters:
        - $ref: '#/components/parameters/ChainId'
        - $ref: '#/components/parameters/Source'
        - $ref: '#/components/parameters/Type'
        - $ref: '#/components/parameters/Status'
        - $ref: '#/components/parameters/RewardTokenAddress'
        - $ref: '#/components/parameters/RewardedTokenAddress'
        - $ref: '#/components/parameters/InvolvedTokenAddress'
      responses:
        '200':
          description: Successfully retrieved incentives
          headers:
            X-Cache:
              description: Indicates if response was served from cache
              schema:
                type: string
                enum: [HIT, MISS]
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetIncentivesResponse'
        '400':
          $ref: '#/components/responses/ValidationErrorResponse'
        '404':
          $ref: '#/components/responses/NotFoundErrorResponse'
        '500':
          $ref: '#/components/responses/InternalErrorResponse'
...
components:
  parameters:
    ChainId:
      name: chainId
      in: query
      description: Filter by chain ID(s). Supports single value or comma-separated list.
      schema:
        type: string
        pattern: '^\d+(,\d+)*$'
      examples:
        single:
          value: '1'
          summary: Single chain (Ethereum)
        multiple:
          value: '1,9745'
          summary: Multiple chains (Ethereum, Plasma)
    Source:
      name: source
      in: query
      description: Filter by incentive source(s). Supports single value or comma-separated list.
      schema:
        type: string
        pattern: '^(ACI_MASIV_API|MERKL_API|ONCHAIN_RPC|LOCAL_CONFIG)(,(ACI_MASIV_API|MERKL_API|ONCHAIN_RPC|LOCAL_CONFIG))*$'
      examples:
        single:
          value: ACI_MASIV_API
          summary: Single source
        multiple:
          value: 'ACI_MASIV_API,MERKL_API'
          summary: Multiple sources
    Type:
      name: type
      in: query
      description: Filter by incentive type(s). Supports single value or comma-separated list.
      schema:
        type: string
        pattern: '^(TOKEN|POINT|POINT_WITHOUT_VALUE)(,(TOKEN|POINT|POINT_WITHOUT_VALUE))*$'
      examples:
        single:
          value: TOKEN
          summary: Only token incentives
        multiple:
          value: 'TOKEN,POINT'
          summary: Token and point incentives
    Status:
      name: status
      in: query
      description: Filter by campaign status. Supports single value or comma-separated list.
      schema:
        type: string
        pattern: '^(LIVE|PAST|SOON)(,(LIVE|PAST|SOON))*$'
      examples:
        single:
          value: LIVE
          summary: Only live campaigns
        multiple:
          value: 'LIVE,SOON'
          summary: Live and upcoming campaigns
    RewardTokenAddress:
      name: rewardTokenAddress
      in: query
      description: Filter by reward token contract address(es). Supports single value or comma-separated list.
      schema:
        $ref: '#/components/schemas/EthereumAddress'
      examples:
        single:
          value: '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2'
          summary: Single reward token
        multiple:
          value: '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2,0x2260FAC5E5542a773Aa44fBCfeDf7C193bc2C599'
          summary: Multiple reward tokens
    RewardedTokenAddress:
      name: rewardedTokenAddress
      in: query
      description: Filter by rewarded token contract address(es). Supports single value or comma-separated list.
      schema:
        $ref: '#/components/schemas/EthereumAddress'
      examples:
        single:
          value: '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2'
          summary: Single token address
        multiple:
          value: '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2,0x2260FAC5E5542a773Aa44fBCfeDf7C193bc2C599'
          summary: Multiple token addresses
    InvolvedTokenAddress:
      name: involvedTokenAddress
      in: query
      description: Filter by involved token contract address(es). Supports single value or comma-separated list.
      schema:
        type: string
        $ref: '#/components/schemas/EthereumAddress'
      examples:
        single:
          value: '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2'
          summary: Single token address
        multiple:
          value: '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2,0x2260FAC5E5542a773Aa44fBCfeDf7C193bc2C599'
          summary: Multiple token addresses
  schemas:
    EthereumAddress:
      type: string
      pattern: '^0x[a-fA-F0-9]{40}$'
    ApiResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
    GetIncentivesResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - data
          properties:
            data:
              type: object
              required:
                - incentives
                - totalCount
                - lastUpdated
              properties:
                incentives:
                  type: array
                  items:
                    $ref: '#/components/schemas/Incentive'
                totalCount:
                  type: integer
                  description: Total number of incentives (before filtering)
                  example: 150
                lastUpdated:
                  type: string
                  format: date-time
                  description: Timestamp of when data was last fetched
                  example: '2025-01-15T10:30:00Z'
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          required:
            - error
          properties:
            success:
              type: boolean
              example: false
            error:
              type: object
              required:
                - message
                - code
              properties:
                message:
                  type: string
                  example: 'Failed to fetch incentives'
                code:
                  type: string
                  example: 'INTERNAL_ERROR'
                details:
                  type: array
                  items:
                    type: object
                    properties:
                      field:
                        type: string
                        example: 'source'
                      message:
                        type: string
                        example: 'Invalid option: expected one of "ACI_MASIV_API"|"MERKL_API"|"ONCHAIN_RPC"|"LOCAL_CONFIG"'
    BaseIncentive:
      type: object
      required:
        - name
        - chainId
        - rewardedToken
        - involvedTokens
        - type
        - source
        - status
        - description
        - claimLink
      properties:
        id:
          type: string
          description: Unique identifier (based on chainId-involvedTokens-rewardTokenAddress|point.name)
        name:
          type: string
          example: 'Supply WETH on Ethereum Core'
        chainId:
          type: integer
          example: 1
        rewardedToken:
          $ref: '#/components/schemas/Token'
        involvedTokens:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/Token'
          description: Non-empty array of tokens being rewarded
        type:
          $ref: '#/components/schemas/IncentiveType'
        source:
          $ref: '#/components/schemas/IncentiveSource'
        status:
          $ref: '#/components/schemas/IncentiveStatus'
        description:
          type: string
          example: 'Earn Etherfi loyalty points on aEthweETH'
        claimLink:
          type: string
          format: uri
          example: 'https://app.merkl.xyz/'
        currentCampaignConfig:
          $ref: '#/components/schemas/CampaignConfig'
        nextCampaignConfig:
          $ref: '#/components/schemas/CampaignConfig'
        allCampaignsConfigs:
          type: array
          items:
            $ref: '#/components/schemas/CampaignConfig'
        infosLink:
          type: string
          format: uri
          example: 'https://governance.aave.com/t/arfc-sts-loop-incentive-program/22368'
        additionalData:
          type: object
          additionalProperties: true
          description: Additional metadata specific to the incentive
    TokenIncentive:
      allOf:
        - $ref: '#/components/schemas/BaseIncentive'
        - type: object
          required:
            - type
            - rewardToken
          properties:
            type:
              type: string
              enum: [TOKEN]
            rewardToken:
              $ref: '#/components/schemas/Token'
            currentApr:
              type: number
              format: float
              example: 5.6
              description: Current Annual Percentage Rate
    PointIncentive:
      allOf:
        - $ref: '#/components/schemas/BaseIncentive'
        - type: object
          required:
            - type
            - point
          properties:
            type:
              type: string
              enum: [POINT]
            point:
              $ref: '#/components/schemas/Point'
            pointValue:
              type: number
              format: float
              example: 1.2
              description: Value of the point
            pointValueUnit:
              type: string
              example: 'per_token'
              description: Unit for point value calculation
    PointWithoutValueIncentive:
      allOf:
        - $ref: '#/components/schemas/BaseIncentive'
        - type: object
          required:
            - type
            - point
          properties:
            type:
              type: string
              enum: [POINT_WITHOUT_VALUE]
            point:
              $ref: '#/components/schemas/Point'
    Incentive:
      oneOf:
        - $ref: '#/components/schemas/TokenIncentive'
        - $ref: '#/components/schemas/PointIncentive'
        - $ref: '#/components/schemas/PointWithoutValueIncentive'
      discriminator:
        propertyName: type
        mapping:
          TOKEN: '#/components/schemas/TokenIncentive'
          POINT: '#/components/schemas/PointIncentive'
          POINT_WITHOUT_VALUE: '#/components/schemas/PointWithoutValueIncentive'
      description: |
        Represents any type of incentive (token, point with value, or point without value).
        The type field determines which properties are available.
    Token:
      type: object
      required:
        - name
        - symbol
        - address
        - chainId
        - decimals
      properties:
        name:
          type: string
          example: 'Wrapped Ether'
        symbol:
          type: string
          example: 'WETH'
        address:
          allOf:
            - $ref: '#/components/schemas/EthereumAddress'
          example: '0xC02aaA39b223FE8D0A0e5C4F27eAD9083C756Cc2'
        chainId:
          type: integer
          example: 1
        decimals:
          type: integer
          example: 18
        price:
          type: number
          format: float
          example: 4250.75
          description: Current price in USD
        priceFeed:
          allOf:
            - $ref: '#/components/schemas/EthereumAddress'
          example: '0x5424384B256154046E9667dDFaaa5e550145215e'
          description: Chainlink price feed address
    Point:
      type: object
      required:
        - name
        - protocol
      properties:
        name:
          type: string
          example: 'Etherfi Loyalty Points'
        protocol:
          type: string
          example: 'Etherfi'
        tgePrice:
          type: number
          format: float
          example: 0.5
          description: TGE price
    CampaignConfig:
      type: object
      required:
        - startTimestamp
      properties:
        startTimestamp:
          type: integer
          format: int64
          example: 1735700400
          description: Unix timestamp for campaign start
        endTimestamp:
          type: integer
          format: int64
          example: 1767236400
          description: Unix timestamp for campaign end (optional, undefined if no end)
        budget:
          type: string
          example: '200000'
          description: Campaign budget (for token incentives)
        apr:
          type: number
          format: float
          example: 8.5
          description: Annual Percentage Rate (for token incentives)
        pointValue:
          type: number
          format: float
          example: 3
          description: Point value (for point incentives)
    IncentiveSource:
      type: string
      enum: ['ACI_MASIV_API', 'MERKL_API', 'ONCHAIN_RPC', 'LOCAL_CONFIG']
      description: |
        - ACI_MASIV_API: Incentives from ACI Masiv infrastructure
        - MERKL_API: Incentives from Merkl infrastructure
        - ONCHAIN_RPC: Incentives fetched directly from onchain data
        - LOCAL_CONFIG: Incentives from local configuration (external points, incentives without end date, etc.)
    IncentiveType:
      type: string
      enum: [TOKEN, POINT, POINT_WITHOUT_VALUE]
      description: |
        - TOKEN: Token incentives (with APR)
        - POINT: Points incentive (with value)
        - POINT_WITHOUT_VALUE: Points incentive (without value)
    IncentiveStatus:
      type: string
      enum: [LIVE, PAST, SOON]
      description: |
        - LIVE: Campaign is currently active
        - PAST: Campaign has ended
        - SOON: Campaign will start in the future
  responses:
    ValidationErrorResponse:
      description: Validation error - Invalid query parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              message: 'Validation failed'
              code: 'VALIDATION_ERROR'
              details:
                - field: 'chainId'
                  message: 'Invalid chain ID format'
    NotFoundErrorResponse:
      description: Not found error - Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              message: 'Resource not found'
              code: 'NOT_FOUND'
    InternalErrorResponse:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              message: 'An unexpected error occurred'
              code: 'INTERNAL_ERROR'